+++ MinecraftJars.Core/Versions/VersionOs.cs
@@ -1,0 +1,8 @@
+﻿namespace MinecraftJars.Core.Versions;
+
+public enum VersionOs
+{
+    None = 0,
+    Windows = 1,
+    Linux = 2
+}