+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Mojang/Os.cs
@@ -1,8 +1,0 @@
-﻿namespace MinecraftJars.Plugin.Mojang;
-
-[Flags]
-public enum Os
-{
-    Windows = 1,
-    Linux = 2
-}